Trait ruukh::component::Component

pub trait Component: 'static {
    type Props;
    type Events;
    type State: Default;
    fn init(
        props: Self::Props, 
        events: Self::Events, 
        status: Status<Self::State>
    ) -> Self;

    fn update(
        &mut self, 
        props: Self::Props, 
        events: Self::Events
    ) -> Option<Self::Props>;

    fn refresh_state(&mut self) -> bool;

    fn status(&self) -> Option<&Rc<RefCell<Status<Self::State>>>>;
}

Trait to define a component. You do not need to implement this trait. Auto implement this trait by using #[component] on a component struct (which also does other magical stuff).

Internals

A component struct which implements this trait usually has the following fields defined into it.

Prop fields - The fields which are used by their parent to pass values into it.

State fields - The fields which are used internally by the component itself to manage its state.

Event field - The field which stores the Self::Events value which are passed by the parent to be invoked when the defined events occur.

Status field: The field which stores the state metadata i.e. stores the mutable state itself, whether state/props are dirty as well as state change notifying mechanism.

Associated Types

type Props

The prop type of a Component.

Internals

The props type is generated by copying all the prop fields from a #[component] struct. This type is named by concatenating component name with Props.

type Events

The event type of a Component.

Internals

The events type is generated by parsing all the event signatures passed in #[events] attribute on a #[component] struct. This type is named by concatenating component name with Events.

type State: Default

The state type of a Component.

Internals

The state type is generated by copying all the state fields from a #[component] struct. This type is named by concatenating component name with State.

Required Methods

fn init(
props: Self::Props,
events: Self::Events,
status: Status<Self::State>
) -> Self

Creates a new component with the props, events and state passed to it.

Internals

It initializes the component with props, events passed to it. The props are used as is whereas the events passed are converted to Self::Events types from the Self::Events::Other type.

It also creates a Default::default() state along with wiring up change notifying mechanism into status.

fn update(
&mut self,
props: Self::Props,
events: Self::Events
) -> Option<Self::Props>

Updates the component with newer props & events and returns older props (if changed).

Internals

When updating the component with newer props, it compares each prop if they changed. Also, it updates the events blindly as their is no point in comparing closures.

fn refresh_state(&mut self) -> bool

Updates the state fields if the status is mutated and returns if actually any changes occured.

fn status(&self) -> Option<&Rc<RefCell<Status<Self::State>>>>

Get the status of the component.